Reorganize and change visibility of struct fields.
- Reorganize into separate files for message and push context types.
- Change visibility of all fields in message and push context to be
  un-exported.

Closes #2.
